UI from @larz0. Refactor from table view to dialog.
