+++ src/htmlContent/extension-manager-dialog.html
@@ -16,7 +16,12 @@
         <div class='spinner large spin'/>
     </div>
     <div class="modal-footer">
-        <button class="btn left install-from-url">{{Strings.INSTALL_FROM_URL}}</button>
+        <span id="install-drop-zone" class="btn left drag">
+            <span class="install-drag-message">{{Strings.INSTALL_EXTENSION_DRAG}} <a class="install-from-url">{{Strings.INSTALL_FROM_URL}}</a></span>
+            <span class="install-drop-message">{{Strings.INSTALL_EXTENSION_DROP}}</span>
+            <span class="install-validating-message">{{Strings.INSTALL_EXTENSION_VALIDATING}}</span>
+            <div id="install-drop-zone-mask"></div>
+        </span>
         <button class="dialog-button btn primary" data-button-id="close">{{Strings.CLOSE}}</button>
     </div>
 </div>
